# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Radhiyan Pet and Care",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/fasilitas": {
            "get": {
                "tags": [
                    "Fasilitas"
                ],
                "summary": "Get Fasilitas",
                "description": "get Fasilitas",
                "operationId": "fasilitas",
                "parameters": [
                    {
                        "name": "body",
                        "in": "path",
                        "required": true,
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "orderby": {
                                            "type": "text",
                                            "example": "asc"
                                        },
                                        "column": {
                                            "type": "text",
                                            "example": "fasilitasName"
                                        },
                                        "keyword": {
                                            "type": "text",
                                            "example": "RPC"
                                        },
                                        "page": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "total_per_page": {
                                            "type": "number",
                                            "example": "5"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Get Data Location Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Get Data Location Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Fasilitas"
                ],
                "summary": "Insert Fasilitas",
                "description": "Insert Fasilitas",
                "operationId": "Insert Fasilitas",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "fasilitasName",
                                    "locationName",
                                    "capacity",
                                    "status",
                                    "introduction",
                                    "description"
                                ],
                                "properties": {
                                    "fasilitasName": {
                                        "type": "text"
                                    },
                                    "locationName": {
                                        "type": "text"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "integer"
                                    },
                                    "introduction": {
                                        "type": "text"
                                    },
                                    "description": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {},
                            "examples": {
                                "Insert Fasilitas": {
                                    "summary": "Insert Fasilitas",
                                    "value": {
                                        "fasilitasName": "Kandang Maxi",
                                        "locationName": "RPC Permata Hijau Pekanbaru",
                                        "capacity": 1,
                                        "status": 1,
                                        "introduction": "Kandang maxi Extra bed for you love pet",
                                        "description": "Ukuran 8M Cocok untuk tipe anjing besar, seperti golden retriever"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Fasilitas Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Fasilitas Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/detailfasilitas": {
            "get": {
                "tags": [
                    "Fasilitas"
                ],
                "summary": "Get Fasilitas Detail",
                "description": "Get Fasilitas Detail",
                "operationId": "detailfasilitas",
                "parameters": [
                    {
                        "name": "body",
                        "in": "path",
                        "required": true,
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "text",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Get Data Location Detail Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Get Data Location Detail Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}